feat(payment): PI-851 moved fields to separate folder
